+++ src-tauri/src/git/types.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 // Note: crate::error::Error will be needed by functions using these types,
 // but the types themselves don't directly depend on it here.
 
-#[derive(Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, specta::Type, Clone, Debug)]
 #[serde(rename_all = "camelCase")]
 pub struct Commit {
     pub sha: String,
@@ -14,7 +14,7 @@ pub struct Commit {
     pub email: String,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, specta::Type, Clone, Debug)]
 pub struct Branch {
     pub name: String,
     #[serde(rename = "fullyMerged")]
@@ -24,7 +24,7 @@ pub struct Branch {
     pub current: bool,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct GitDirResponse {
     pub path: String,
@@ -32,18 +32,18 @@ pub struct GitDirResponse {
     #[serde(rename = "currentBranch")]
     pub current_branch: String,
     #[serde(rename = "branchesCount")]
-    pub branches_count: usize,
+    pub branches_count: u32,
     pub name: String,
     pub id: String,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone)]
+#[derive(Serialize, Deserialize, specta::Type, Debug, Clone)]
 pub struct DeletedBranchInfo {
     pub branch: Branch,
     pub raw_output: String,
 }
 
-#[derive(Serialize, Deserialize, Clone, Debug)]
+#[derive(Serialize, Deserialize, specta::Type, Clone, Debug)]
 #[serde(rename_all = "PascalCase")]
 pub enum ConflictResolution {
     Overwrite,
@@ -51,7 +51,7 @@ pub enum ConflictResolution {
     Skip,
 }
 
-#[derive(Serialize, Deserialize, Clone)]
+#[derive(Serialize, Deserialize, specta::Type, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct RestoreBranchInput {
     pub original_name: String,
@@ -60,7 +60,7 @@ pub struct RestoreBranchInput {
     pub conflict_resolution: Option<ConflictResolution>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, specta::Type)]
 #[serde(rename_all = "camelCase")]
 pub struct RestoreBranchResult {
     pub success: bool,
@@ -72,7 +72,7 @@ pub struct RestoreBranchResult {
     pub branch: Option<Branch>,
 }
 
-#[derive(Serialize, Deserialize)]
+#[derive(Serialize, Deserialize, specta::Type)]
 pub struct ConflictDetails {
     pub original_name: String,
     pub conflicting_name: String,
